RemoteBatchPropertiesWithFilter, fix bug in property access on parameters ()
### Part 1:
Fixes a bug in `RemoteBatchPropertiesWithFilter` where a predicate
expression containing a parameter with properties would try to look up
the parameter name from the slots.

For this query
> MATCH (n)
> WHERE n.prop = #param.prop
> RETURN n.prop

the right side of the predicate has the parameter wrapped in a
`Property` expression

>  Equals(
>     Property(Variable(n),PropertyKeyName(prop)),
>
Property(ExplicitParameter(param,Any,UnknownSize),PropertyKeyName(prop))
>  )

When we extract the predicate variable this Property expression would
try to look up the slot offset for the parameter name, which throws an
exception since the parameter doesn't exist in the slot configuration.
This fix will skip any Property that has a parameter as subexpressions
when extracting predicate variables.

### Part 2:
Fix transactions in Scala spd acceptance tests not running with
`TransactionalContext.DatabaseMode.SHARDED`. The acceptance tests were
using a `TransactionContextFactory` where `DatabaseMode` was hardcoded
to `DatabaseMode.SINGLE` which means that queries were not planned with
`RemoteBatchProperties`.
Now the test context factory will get the database mode from the
`EnterpriseGraphDatabaseTestSupport.runOnSpd` flag that is used to
configure tests for spd.

The `SpdReadPropertyRuleEnforcementAcceptanceTest` extends a class with
test cases that use the regular `"CREATE DATABASE ..."` to setup
multiple dbs, this creates a mix of sharded and single instance dbs as
the default database is a sharded db. Those test cases have been moved
into the non-spd configured test for now, probably they should be
updated to also support creating spd databases.