DWYW: Typed wrapper in codegen ()
I think I have arrived at a way to address some of the issues I have
encountered when writing and reading generated code:

1. lack of static typing causes higher mental load for tracking types
2. when reading generated code it can be hard to grasp what concerns IR
("template-time") vs regular runtime/compile-time scala code
3. IR can look dissimilar to the end result which makes it harder to
read than it needs to be. we end up writing comments to document what
the generated code will look like, which can get stale

The proposal in this PR takes an "incremental typing" approach similar
to TypeScript: add a wrapper type that augments a value with additional
type information. This means there does not need to be any kind of
big-bang migration, and it is fully compatible with the existing
weakly-typed code.

The typed wrapper goes some way to solving (1): variable and field
assignment can be typechecked. Method calls are slightly more tricky
because of the heterogeneous parameter types but I think they are
doable, I will try something out if this PR is accepted.

While adding the new typed API it became clear that I should not modify
any of the existing `IntermediateRepresentation` structures or helper
methods, because of the volume of existing code that uses them: any
small change causes cascading changes throughout the code. On the other
hand I wanted to stay close to the original API to make it easier to
migrate gradually. I also thought it would be helpful to visually
differentiate between untyped and typed IR, so decided to use the `#`
prefix for strongly-typed symbols. This provides a way to mitigate (2),
since any code written with this convention can be more easily
understood as "template-time".

Lastly I have added a few extension classes for particular wrapped types
like fields, variables and methods, which have relevant
methods/operators for those types which look closer to the real
generated code - for example, `setField(field, value)` becomes `#field
:= value`. This helps with (3)