Support vector types in Range indexes ()
This adds support for vector types in GenericKey, by implementing a new
Type (VectorKeyType). Notable discrepancies compared to other types are
the following:

* The type does not include concrete extreme values since it is
unnecessary to allocate a ton of data for searches when we know the
property of the extreme value.
   
* The type uses a short-form for toString, to avoid the creation of a temporary new
Value.

* The type compares its byteArray component-wise in the coordinate type.

The commit also adds a configuration to RandomSupport to limit the
dimension of vectors by their physical size, and reworks its configuration
make it easier to compose.