Makes sure the root layer of the MultiRootGBPTree is never versioned ()
The root layer is like the internal nodes, in that they it should not be
versioned, only the key value pairs in the leafs should be versioned.
This makes sure all readers can see all values, and then MVCC mechanisms
determine if we should be able to read them or if they are not visible
to us.