Termination of remote transactions is an asynchronous operation now
Historically, transaction termination has been a quick and non-blocking
operation and, unfortunately, users count on that (Bolt server invokes
it on Netty event loop thread).
Because of that let's invoke the termination of remote transactions
asynchronously in a fire and forget manner. We don't care about the
result anyway. It is important just to try to terminate remote
transactions using the best effort.