Add heartbeating to prepare store copy ()
This PR addresses two issues:

Currently when a store copy gets initialized with `PrepareStoreCopyRequest` the netty thread is used to take the checkpoint. This is not ideal.
If checkpointing takes longer than the inactivity timeout (default: 10 minutes) then the client will close the connection. The checkpointing will continue and a subsequent `PrepareStoreCopyRequest` will be able to make use the previously started checkpoint; so if the client repeats its request it is just fine. Expect the ugly messages in the logs. Now cluster internal store copy requests will be repeated for sure, but backup related ones won't - unless those are taken by some automation, which won't take no for an answer.

**Part 1**
In this PR the checkpointing is moved to a different thread. Once the request arrives this thread will start the checkpointing, and the netty thread is given back to the pool. Once the background thread has completed and has finished with checkpointing it will hold on the checkpoint lock, while a netty thread will respond with the unrecoverable files and the `PrepareStoreCopyResponse`. The lock is released once we have confirmation that the response was sent.

**Part 2**
Beside the thread which does the checkpointing a second one is started which will just wait for the checkpoint be done OR half the inactivity timeout has passed. This is done with a `CountDownLatch`, which supports waiting just for a specified time. If the checkpointing has finished, this thread just finishes. If not, then it will send a `Heartbeat` message to the client; then it starts to wait again. This `Heartbeat` resets the inactivity timer of the client. In case the channel was closed (e.g. the client cancelled its request), it can't send anything and it won't repeat waiting.
Since this second thread is using a new protocol element it is only started with version 8.