DoC: LogicalPlan2PlanDescriptionTest  refactoring ()
Split LogicalPlan2PlanDescriptionTest as it has gotten very large, now
there's one for:
- regular Cypher: QueryLogicalPlan2PlanDescriptionTest
- admin commands: AdminLogicalPlan2PlanDescriptionTest
- index, constraint, graph types, non-admin show/terminate commands:
SchemaAndNonAdminCommandsLogicalPlan2PlanDescriptionTest and a test base
to hold the common methods and variables

Also sets the effective cardinality in the tests to 1 for index,
constraint and graph type command tests.
Since it isn't really relevant for these kinds of plans, they have a
one-to-one ast-plan mapping (this follows how the admin command plan
tests looks).