Add vectorization provider and support implementation for GQL vector distance and norm functions ()
The following takes inspiration from the approach of many libraries.
Both Lucene and JVector use the same pattern.

All vector operations go through `VectorUtil`, which handles
the precondition checking and dispatches the request to a
`VectorUtilSupport` implementation, the workhorse.

The default implementation uses simple loops; though could be updated
to use manual unrolling and FMA for performance. Though a little
convoluted, this pattern allows us to also make a provider for the new
vectorization API which is in incubator and use that if available.
The `VectorizationProvider` interface will handle detection and
initalisation of the available functionality.

The GQL vector distance and norm functions have been implemented as they
currently are in the spec.