fix(Semantic Analysis): Subquery expressions report correct dependencies ()
For `ScopeExpression`s we need to calculate the dependencies based on
the definitions in the symbol table in
`computeDependenciesForExpressions`:

However, the position is that of where the variable was originally
defined and not where it was used within the subquery expression. Also,
we would in a lot of cases report dependencies on more variables from
the outer scope than were necessary.