Implementation of empty store seeding, to be used in clustering ()
Uses an importer to create empty store, zipping it up into a byte[] for
use in distributed database creation in clustering. Also works for SPD
where the seed will contain all shards, and restoring from such a seed
cherry-picks only the relevant shard from it.

There are some changes to `SingleFilePageSwapper`, allowing it to be
told not to do actual file force calls (for much better performance
running the importer which generates the empty store). Also
`TransactionLogInitializer` now accepts a `Config` so that e.g. tx log
file pre-allocation can be disabled.