SeekCursor remove concurrentWriteHappened field ()
Another refactoring in the SeekCursor.
One of the difficulties in navigating `SeekCursor` code caused by hard
to follow flow of data through the fields.
In this PR I target `concurrentWriteHappened` field which is retry flag
used to ensure we read stable version of leaf.

Whenever we exit `next()` `concurrentWriteHappened` is false, so this
flag is local to the `next()` method.
And on the first call to `next()` we can use field `first` to detect if
it the first call.

See commits one by one to see how I move assignments to
`concurrentWriteHappened` up to emphasise it's limits and do other small
refactoring steps.