Introduce additional memory reservation on deletion operations ()
In MVCC, transactional work is processed in chunks of a configured size.
Once this size is reached, the chunk is formed and processed. However,
the commands representing the actual changes are generated only after
chunk formation. This can lead to significant memory usage growth during
commands generation (especially in cases involving massive deletions)
due to the large size of the resulting command set. Additional memory is
reserved as part of each deletion to partially mitigate this and
decrease the original change set of each chunk. This should allow chunks
to be formed more quickly and keep command generation memory consumption
lower.

While this solution isn't perfect for all scenarios, we believe it will
handle the majority of cases effectively.