Upgrader mechanism ()
In this PR, we implement an upgrade mechanism scheduler
(`RaftUpgraderScheduler`) by wiring this check for the upgrade
versioning into the Raft lifecycle.
There are two schedulers: a concrete one (`RaftUpgraderTaskScheduler`)
and an empty one (`DisabledRaftUpgraderScheduler`) in case the upgrade
via Raft is disabled (via
`internal.dbms.use_clustering_upgrade_mechanism`).

Apart from the tests related to the scheduling and testing the lifecycle
ordering, in this pr there is a small refactor involving moving around
some files from `com.neo4j.causalclustering.core.state.machines.upgrade`
to `com.neo4j.causalclustering.core.upgrade` which are not related just
to the upgrade state machine, but more generally to the upgrade package.