Default index provider based on kernel version ()
During a rolling upgrade, we have to respect the current version of the database to make sure that we don't start using an index provider that is only available in the newer version. This is important since there might still be members using the old binaries that will panic if they receive an index definition from an unknown index provider.