added missing name and description metadata to neo4j-server-test-utils pom ()
The publish to maven central failed because this module was missing the
`<name>` and `<description>` metadata.
This PR adds the metadata to the pom.

I tried experimenting with the maven enforcer plugin to force all poms
to have the name/description metadata in the future but I couldn't get
it working.

I don't know why it suddenly started failing this pom, since it
obviously hasn't been touched in years. I can only assume that maven
central have changed their validation rules.