More fixes for compat spd test fails ()
Fixes for `SpdBlockImportIT` that missed a config setting to enable spd

Disabled tests for CDC since that isn't supported in spd at the moment.

Also disabled some tests for indexes that either asserted on very
specific counts that isn't correct for spd or that used the dependency
resolver to read directly from the graph shard, which will give the
wrong results since it has no properties.

Fixed so that the cypher generator can handle range predicates that
includes everything, solves test failures in
`FindEntityByTokenAndPropertyIT`

[Spd compat build](
https://live.neo4j-build.io/buildConfiguration/BuildServicePipelines_Neo4jSpdIntegrationTests/25216572)