feat: add support for transaction type in Query API ()
This update brings support for transaction type in Query API when
beginning a transaction. This is similar to the existing support in Bolt
Server that accepts `tx_type` field in `BEGIN` message. This is for
internal use and is needed when an unrestricted `IMPLICIT` transaction
needs to be initiated for JDBC autocommit mode.