Add uniqueness constraint on displayName to system db. ()
Part of graph reference work. In the edge case that a composite user has
conflicting displayNames e.g. a database or alias 'foo.bar' and simultaneously a constituent
'foo.bar' in a composite 'foo', the constraint cannot be created on
upgrade. In this case, the upgrade should fail gracefully with an
informative error message.

---------

Co-authored-by: Balazs Lendvai <balazs.lendvai@neotechnology.com>