Provide the metadata script in write support ()
With the `existingMetadata` option the `CREATE DATABASE` command can be instructed that it should execute the metadata script contained by the header of a backup archive. It is only available if the database gets seeded with the `seedUri` option.

This `seedUri` used to be not checked while the running the `CREATE DATABASE` command, only once the servers started the database and tried to use the value. Any problem with the `seedUri` was only communicated with the databases not starting and providing some error message (sometimes very hard to understand) in `SHOW DATABASES`.

The `seedUri` is now extensively checked in the write support (at a preparation step before the `CreateDatabasePlanner` actually creates the `:Database` nodes and related entities). Checking means, the backup header is checked, while doing this the metadata script is automatically available, meaning the metadata script is available one the write support returns.
Also all problems with it (the backup archive not having it, pointing to a dump, the archive not existing, etc.) are handled there. In case of the `seedUri` pointing to a directory (with backups in it) the metadata script will be retrieved too (this was not possible before). 

This means the rather convoluted way of using a `BackupMetadataScriptProvider` becomes unnecessary. The metadata script (if available) just needs to be transferred between the `CreateDatabasePlanner` and `ApplyMetadataScriptIfSuppliedPlan`. Latter then becomes much simpler as it does not need to have code for acquiring the script.
Transferring the script was quite a challenge for me: it gets transferred via context parameters with `UpdateContextParams` in `QueryHandler.onResult`.

Since the `BackupMetadataScriptProvider` component is removed so is the complication with the timeouts and futures in it. I don't see why that was really needed. The call to the write support can be held up indefinitely already (as it takes and wait for a lock on all topology graph operations) and that was not considered a big enough problem to act on it.

This PR does not do much for sharded databases - metadata scripts for those are not available yet on the provider side. No backup fetches the metadata script for the virtual node, and the scripts for individual shards are useless. For sharded databases further changes are needed in several places: backup, metadata script generation, shard seeds checking, etc.