Improve clustered kernel version upgrade logging ()
The
`TransactionLogRotationOnBackupAndCopyIT.rotatingToNewFileOnVersionChangeForBackups()`
test failed during KernelVersion/LogFormat upgrade on one of the cluster
followers. Somehow managing to rotate the transaction log file, but
remain on the older LogFormat, which in turn broke checksum consistency
of the cluster. I was unable to identify a conclusive mechanism and
couldn't reproduce the failure locally. However, diagnosis was also
hampered by the very limited logging on the processing of the version
upgrade, especially on the cluster followers.

This PR fixes a few minor places where possibly an older LogFormat could
be used during rotation, but mostly adds some logging of
KernelVersion/LogFormat at database start and when the UpgradeCommand is
applied to give more information if this failure reoccurs in the future.