Show VirtualSPD only when getting aliases ()
Resolve some errors with how `SHOW DATABASES` and `SHOW ALIASES` were
getting the database references such that it displayed as if the alias
was on the graph shard instead of the virtual database (changes to
`DatabaseListParameterTransformerFunction.scala` and
`ShowAliasesExecutionPlanner.scala`.

Also update alias acceptance tests so that when `runOnSpd=true` is set
these pass.