Remove possible race-condition in storecopy mutex ()
This change removes the possibility of the periodic checkpointer taking
a write lock between the prepare storecopy process releasing its write
lock and acquiring a read lock.

When prepare store copy checkpoints, it gets a reference to the read
lock `a` then attempts to checkpoint, acquiring a write lock `b` while
the checkpoint is running, the periodic checkpointer tries to acquire a
further write lock `c` to begin checkpointing.

the locks are now `b` is the active mutex, `c` is at the front of the
queue, and `a` is unqueued, when storecopy checkpoint completes it
releases `b` and `c` is able to acquire the write lock, `a` is then
queued up behind a further checkpoint. But read lock `a` is required to
send the prepare store copy response.

cl: Fixed a bug that could cause backup requests to take longer.