Initial delay for BufferedIdController#maintenance ()
This deflakes
MultiVersionDatabaseRestartIT#writerWithReallyOldVersionAfterRestart.

A previous attempt to deflake the test set the
maintenance interval to 5h, overlooking that the
background job still runs the first time directly
when scheduled. When debugging it could be seen that a concurrent ID
maintenance is the cause of the failing test. This makes it so that the
background job for ID maintenance waits
`GraphDatabaseInternalSettings.id_controller_maintenance_interval`
before running the first time. In practise, when the setting is set to
one second, there should be no observable behavior difference.