Fixes a bug where logformat would not be correctly chosen on upgrade
Not a real problem yet since no new log format has been introduced.

If upgrade transaction was in a batch with another transaction (after
it), the logformatprovider would not be updated at the point of rotation and the
wrong format selected.
For kernel version this could be fixed in TransactionLogWriter.append by
checking the version of the tx,
but for logformat we don't have access to what format to choose.
Instead, it will be up to the logic batching transactions to never add
anything to a batch that contains upgrade already.

This fixes the replicated tx batching process and the read replica batching process