SPD: Making some operations on property shard transactions asynchronous
A graph shard will not wait for the result when committing or rolling
back a property shard transaction. This means a performance improvement
for short transactions that fetch only very little data from property shards.
Why is it OK not to wait for a result of commit or rollback operations?
This path is used only for reading, so the only effect of commiting a
transaction is freeing resources both on the related graph shard and the
property shard. So it is OK not to wait for the result of this resource-freeing
operation.

There is an internal setting that can disable this behaviour.