Fix to not create a circular GQL cause ()
We have seen some instances where a circular cause gql-errors has been
created by mistake, resulting in stack overflow errors instead.
This introduces a check so that we don't add a cause more than once to
the cause chain, so that if we create the errors wrong in the future we
won't get stack overflow.