Introduce a setting to control log format upgrade
A setting that will allow upgrading to the next log format on any
following upgrade, and control the format of new databases.
Introduced as a way to turn on a new log format for some but not all
customers in a particular binary version.
Disconnects tx writing decisions that were tied directly to kernel version
to make decisions based on log channels/format.

Additional step to follow, for now it can only control it on GLOURIOUS_FUTURE.

Also adds an extra setting temporarily to be able to test with/without our
envelope profile