Add GQL status and neo4j status for vectors not supported in record storage ()
Motivation for the "db-format-2000" spread around tests:
* There are many tests that need to pass in that store format String,
and it is only used in one place so far.
* Figuring out, for each test, where to get the correct store format
name from would take some time and is not really necessary.
* Should the usage pattern change in the future, an error with
"db-format-2000" makes it easier to locate the origin of that String
than "block" or "aligned".

(Happy to discuss if that does not make sense to you.)