test(LogicalPlanBuilder): Fix variable escaping in LogicalPlanToPlanBuilderString ()
Align LogicalPlanToPlanBuilderString with AbstractLogicalPlanBuilder wrt
escaping.

We need to escape variables in places where expressions can be used or
when used in pattern.
However, we also escape in some other cases like the leaf operators.

This should probably be fixed at some point but this test makes sure
that we at least are aligned and can copy the logicalPlanBuilderString
to build a logical plan