Expand RandomValues configuration ()
During debugging, I found it useful to add some ways to further restrict
which types are allowed to be generated, and to further limit the
allowed vector dimensionality.

* Switches to a builder pattern to make a record from creating multiple
intermediate records.
This allows for the invariants to be checked and not broken between the
builder-methods, and fewer intermediate heap objects (which cannot be
elided/inlined in
  Java  [to my knowledge])
  
* Adds new methods in the configuration for limiting the allowed types
and the dimensionality. Some override other concepts, so this is
reflected.

* Adds convenience methods for min/max  builder-methods.
  
* General cleanup of naming of enums to be more consistent with
  elsewhere, and to use the enhanced switch syntax.

Overall, the change isn't huge, but the propagation of using a builder
touches a lot of files.