Add tracing numbers in tests for spd ()
Added tracing numbers for spd in tests, since record sizes are smaller
than normal block on the entity graph in spd we should get a bit fewer
pins in spd compared to block. Setting the sizes the same we got the
same number of pins for spd and block.

For the tests that are skipped on spd right now I commented out other
assertions in the test and then ran it to check the spd values.

For `TransactionTracingIT.tracePageCacheAccessOnAllNodesAccess` the
tracing check was missed before so added that back in and then I had to
change from `pins()` to `pins().atMost()` because the number of pins is
always one more than the number of unpins for all formats in this test.
Looking back at the history of the test I can see that we asserted on
the separate values being different before in this test so seems fine to
do this change, and better than not having any assertions at all.