Block format: Fixes bug in schema cache "no such constraint"  ()
Block format uses the order of the schema changes provided from the tx
state when applying transactions.

In record format the schema changes are deliberately reordered so that
the schema changes are Deletes first Creates second.

This commit reorders the schema changes in the tx state so that block
has the same behaviour as Record format.

Included is a test that before these changes would fail (for block
format) due to an invalid in between state of the schema cache where the
delete happening after the create caused the accidental half removal of
the newly created constraint from the schema cache

I also snuck in some variable renames for more clarity in the tests