fix(SPD Compatability): Index Seek With Composite Filters ()
* When using composite filters (i.e., either multiple predicates on the
same property, or multiple predicates using a composite index), unless
all the predicates in the selection support index seeks, only the
largest prefix of predicates that support index seek are pushed to the
index for filtering and the remaining are then applied on the rows
retrieved from the index. We were earlier incorrectly assuming that all
these filters were applied on the index directly, which resulted in an
unnecessary remoteBatchProperties instead of fetching the properties
from the index itself. This commit fixes this by looking for all
predicates that have been replaced by existence predicates and ensuring
that the properties used there are fetched from the index.

* In the nodeIndexSeekAcceptanceTest we noticed we were not
pre-filtering on selections before running remoteBatchProperties in the
slotted/interpreted runtimes (RemoteBatchPropertiesWithFilter is not
supported on these runtimes). This is now fixed in the
ShardPredicatePushdownPartition.