Graph types - initial runtime parts ()
This solution is dropping everything and then creating the asked for things.
This means that existing things we want to keep will be dropped and re-created.
This will later on be updated to not drop and re-create things we want to keep.

This PR also splits the SchemaCommandRuntime into one enterprise and one community version, leaving the enterprise create constraint commands in community to not change their existing errors on not being available in community to the new one used by the graph type command.

Co-authored-by: Mark Dixon <1756429+mnd999@users.noreply.github.com>