Create VectorStores on migration ()
When migrating within block format from a version before vector stores
to after, the store would not exist. The migration would bump to latest
kernel version. This would fail one of the first checks in recovery,
as it is expecting that file to exist from that version. Migration
therefore needs to create that file. To ensure that it occurs for all
paths, it is created in the post-migration step.