Enveloped directPutAll adds padding if necessary
Fix EnvelopeWriteChannel.directPutAll to handle padding that occurs
lazily on the source writer.

If the sender of the raw bytes send tx 1-3, and then send tx 4-5 later,
any padding between 3 and 4 would be missed. Now the receiver adds
padding if it should be before the received rawbytes.

Also adds an extra rotation for enveloped logs on receiver side for an
rotation on sender side instead of just rotating on rotation limit.
Might not be necessary for production but test that do manual rotation
were seeing problems.

Co-authored-by: Emma Holmberg Ohlsson <emma.holmberg.ohlsson@neo4j.com>