Fix handling of upgrade transaction in the RTSM ()
This is a fix for a problem in the `ReplicatedTransactionStateMachine`
when handling a stream of transactions with different kernel versions.

**Problem** 
When we call `applyCommand` in the state machine, we roughly do this: 
```
applyCommand(commandIndex,txAppliedCallback): 
  tx.setConsensusIndex(commandIndex);
  registerListener(commandIndex, txAppliedCallback); 
  queue.add(tx);
```

The problem arises within the listener. Consider an explicit scenario,
with a clean slate, in which we start by applying the upgrade
transaction, which itself has the old kernel version:

```
applyCommand(commandIndex=0, kernelVersion=OLD):
   # Which creates this state:  
   listenerState = {listeners=[0], batchStartIndex=0} and queue=[0]
```

Then comes a new transaction, after the upgrade, with a new kernel
version:
```
applyCommand(commandIndex=1, kernelVersion=NEW):  
  # First, we call `registerListener`  after which the state is: 
  listenerState = {listeners=[0,1], batchStartIndex=0} and queue=[0]
  ...
```

Now, still inside the `applyCommand`, we will call `queue.add(tx=1)`.
Since this
[change](https://github.com/fabiim/neo4j/commit/1ba212f0ccf908e3d913406689072d5689544238)
that will trigger a flush of the queue state before queing the new
transaction because the kernel version is different from the tail of the
queue. When we flush and apply `0` we will call
`ReplicatedTransactionStateListeners#notifyListeners(tx=0)`. **That will
callback all the listeners it has (`[0,1]`)** and then clear the
listener state. Finally, we queue transaction `1`. So by the end of
applying transaction `1` we have this state:

```
applyCommand(commandIndex=1, kernelVersion=NEW):  
   listenerState = {listeners=[], batchStartIndex=null} and queue=[1]
```
**This is BAD because we triggered the callback on transaction `1` that
releases it before it has been applied. **

Now we process the next transaction: 

```
applyCommand(commandIndex=2, kernelVersion=NEW):  
   listenerState = {listeners=[2], batchStartIndex=2} and queue=[1,2]
```
The problematic state that has led us to find this bug has been created,
but it will only cause issues when we flush the queue, which can happen
much later after sufficient transactions are added to the queue or we
explicitly flush it.

Once we flush the `queue` and apply the contained transactions, we start
by applying transaction `1` and that will invoke the
`notifyListeners(tx=1)`, which does this validation and fail:

```
 CommandBatch commandBatch = transaction.commandBatch();
        if (commandBatch.consensusIndex() != batchStartIndex) {
            throw new IllegalStateException(format(
                    "Expected events for command batch starting at index %s, but received from batch starting at index %s.",
                    batchStartIndex, commandBatch.consensusIndex()));
        }
```

Because `commandBatch.consensusIndex=1` and `batchStartIndex=2`. 

**Fix**

We need to inspect the state of the queue to determine if it will queue
the transaction with the existing transactions in the queue or not. That
is because the listeners and the transaction queue are tightly coupled;
once we register a callback on the listeners, it is assumed that they
will be called after the current transaction queue is flushed.