Update GenAI plugin ()
Many models have been deprecated in VertexAI. The plugin did not follow
the changes, and all "supported" models have been retired. This means
that they cannot be used; their endpoints have been disabled.

The restriction of the models was originally due to the mitigation of
security concerns, as they modified the endpoint URI of the request.
Since then, other areas of the plugin have been allowed to be relaxed in
this regard. The area of the endpoint URI that is modified by the user
input of the model is past the point where it could hijack the URI and
redirect it elsewhere.

This removes that restriction in both VertexAI and Bedrock.

One of the models for VertexAI can only accept single resources per
query; thus it was needed to separate batched and single encoders, where
the single handles the resources in their own requests, and the batch
can send them together.

However, despite being deprecated and removed, it would be a breaking
change to change the default model for VertexAI; the model is different,
and silently making semantically incomparable vectors could potentially
introduce semantic inconsistencies in returned results. For the time
being, the existing default will throw as if it is an unsupported model.

Whilst here, we update the allowed regions and add known new parameters.

In the case of VertexAI, we can safely add these parameters, as all
previous models supported have been deprecated, and there is no risk of
a model rejecting the request.

Bedrock still supports the existing Titan G1 text embedding model and
rejects the parameters, whereas the multimedia embedding model
configures dimensions differently. Bedrock now has an encoder per model,
with a fallback of no additional configuration.