Adapt text-index tests to work for spd compatibility ()
Since we use Cypher to communicate with the shards the index that we use
may differ compared to single instance if the plan for one is better.

When looking at `leafPlanOptions.tieBreaker` we see that if the cost for
two plans is the same we look at other factors to see which might be
better and then in `leafPlanOptions.indexTypeModifier` we only use text
for `ENDS WITH (= SUFFIX)` and `CONTAINS`. So for `STARTS WITH
(=PREFIX)` the range index is better to use. Therefore the assertions in
the test no longer holds.
I think that the tests have value to run on SPD as well so added a spd
version, could not update the test in an easy way since it's a community
test and we don't have access to spd code.