Include name in equals in ConstraintDescriptor implementers ()
This is something we've previously decided was too cumbersome with no
added benefit besides cleaner code. But! now we found a bug that get's
solved with this change. Also updates
`HashCode()` of constraint descriptor implementors

The places that relies on the previous equals functionalities for ConstraintDescriptor implementers now instead use the aptly named equalsIgnoreName, which does the same thing as the old equals.